Stack-based shared memory allocator (#703)
This PR replaces `NoReuseSharedMemAllocator` with a new shared mem
allocator called `StackBasedSharedMemAllocator`. In this new approach,
we re-use memory that is no longer used where there are existing
synchronizations in the kernel. In particular, we use a stack of
allocations to lay out memory, and we delay pushing onto the stack until
it is time to reclaim memory. This lets us order the allocations not by
when they were first written, but by when they are last read, giving us
more opportunities for re-use since the first allocations to be freed
will lie at the top of the stack.

In a future PR, we can explore extensions of this approach which
introduce new syncs. Since the current implementation does not introduce
syncs and only reuses memory when it is safe to do so, it is activated
as the only method for shared memory allocation, replacing the previous
class that never reused space.