Remove `resolveIndexDtype` (#713)
I changed `resolveIndexDtype` into:

```C++
void Val::resolveIndexDtype() {
  TORCH_INTERNAL_ASSERT(dtype_ == DataType::Index);
}
```

And run all C++ tests and benchmarks. The only failure is
`TensorFactoryTest.StandaloneARange`. I think it is fine, because the
"index" of `arange` by definition can be an arbitrary dtype including
double and complex, etc., so having an explicit int64 or int32 will not
be a problem.