More smartly determine `MAX_JOBS` (#741)
The default `MAX_JOBS` should make the build as fast as possible, while
still run out of memory. I tested locally with gcc 13.1, with 64 jobs in
parallel, the highest memory usage is 181GB, so roughly 2.8GB/task. To
be safer, we should consider capping the max jobs to `available mem /
3`.

---------

Co-authored-by: jjsjann123 <alex.jann2012@gmail.com>