Don't count parallelized loops in AllocationInfoMap
Previously, we stacked every ForLoop regardless of parallelization. This
meant that when the first few dimensions were left of compute at in the
whole fusion, even if they were parallelized all tensors would have the
same outer live interval. I noticed this for the
AmpereMatmulSmemEpilogue_CUDA tests. In that case if you look at the
generated CUDA it's clearly not true; the outer for loops do not appear
since they are parallelized. This commit fixes this; note that it can
affect all reuse analysis including aliasing even of local memory.