Refactor `Opaque` (#756)
Move it to a separate file. Also, the equals operator is now handled
inside the ctor of `Opaque`, so that there is no possibility of
inconsistency. In the future, this change also makes `Opaque` more
extensible, for example, if I want to get the vector of bytes for the
data in `Opaque`. I believe this new approach is clearly better than the
pervious one, and I have no idea why I wasn't doing it like this at that
time.