Add `TensorView::promoteReuse` (#739)
As of #703, nvfuser is able to reuse shared memory even when the indexes
don't match. However, this requires block synchronization to occur in
the appropriate places. That PR did not provide a mechanism for
inserting synchronization. This PR addresses this by adding the method
`TensorView::promoteReuse()`, which sets the `promote_reuse_` flag on
the tensor. The `reuseMemoryAllocations` lowering pass recognizes that
flag and ensures synchronizations are available after the tensor's last
use but before the next smem allocation is written to, so that memory
reuse can occur. That pass now looks like:

1. Find shared or local memory tensors that can be "aliased". That is,
their lifetimes don't overlap and they have the equivalent index
expressions, so the later one can be replaced with a reference to the
first.
2. Find shared memory tensors which we should promote for re-use, i.e.
those with the `promote_reuse_` flag set. These determine intervals
between their last read and the next first write of another smem tensor;
we check for syncing expressions within those intervals. Currently we
insert a sync at the end of the interval if we don't find a pre-existing
one, but we could change these to arrive/wait barriers in future work.
3. Do shared memory allocation as introduced in #703. The new syncs
introduced in step 2 are now recognized and memory is reclaimed as
requested.

Note that currently `promoteReuse` can be called on any `TensorView`,
but it only has an effect on shared memory tensors.