Randomize testing with NVFUSER_TEST_RANDOM_SEED
This makes tests random but reproducible. Instead of always setting
random seeds for ATen and `std::srand` to 0, we set them to the
current time by default. If an error in a test occurs, we print the seed
that was used after the backtrace. To force another seed, use the
environment variable `NVFUSER_TEST_RANDOM_SEED`.