Stop extending lifetimes past end of for loops in aliasing pass (#774)
Currently, `ScopeMap` increments the position when recording the end of
either a `ForLoop` or the top-level scope. This means we cannot
interpret the end position as an actual last read, breaking the
guarantee that `BufferLiveInterval` is a closed interval. Since we do
not have separate expressions for the end of loops, we cannot point to a
specific location for the loop ending, but interpreting the interval as
half-open depending on the loop nest structure is confusing, as seen in
#772.

This PR introduces positions for the ends of for loops, although they do
not have corresponding `Expr`s in the kernel IR. As a result, two
adjacent for loops will have non-intersecting intervals, and allocations
with uses confined to each will thus have non-intersecting outer live
intervals. Since the intervals are closed, if an actual expression is
the last read of one interval and first write of another, they will be
properly considered as overlapping, and not eligible for inserting a
sync.

The example from #772 now looks like:
```
1  T0_s[ iS0{7} ] = ALLOCATE(buffer=T0_s[ iS0{7} ], mem_type=shared, size=7, zero_init=false)
   ^^^^^ ---Buffer Reuse Info---   not aliased  , inner live interval: [ 7 , 11 ] , size expr : 7 , outer live interval: [ 6 , 12 ]    
2  T2_s[ iS3{8}rf ] = ALLOCATE(buffer=T2_s[ iS3{8}rf ], mem_type=shared, size=8, zero_init=false)
   ^^^^^ ---Buffer Reuse Info---   not aliased  , inner live interval: [ 14 , 18 ] , size expr : 8 , outer live interval: [ 13 , 19 ]    
3  T4_s[ iS6{9}rf ] = ALLOCATE(buffer=T4_s[ iS6{9}rf ], mem_type=shared, size=9, zero_init=false)
   ^^^^^ ---Buffer Reuse Info---   not aliased  , inner live interval: [ 21 , 24 ] , size expr : 9 , outer live interval: [ 20 , 25 ]    
4  f2 = ALLOCATE(buffer=f2, mem_type=register, size=1, zero_init=false)
5  f2 = (float)(1);
6  FOR i58 in iS0{7}:
7    T0_s[ iS0{7} ]
   = full({7}, ( (float)(1) ));
9  T1_l[ iS1{7} ] = ALLOCATE(buffer=T1_l[ iS1{7} ], mem_type=register, size=7, zero_init=false)
   ^^^^^ ---Buffer Reuse Info---   not aliased  , inner live interval: [ 11 , 14 ] , size expr : 7 , outer live interval: [ 10 , 15 ]    
10  FOR i59 in iS1{7}:
11    T1_l[ iS1{7} ]
   = -T0_s[ iS0{7} ];
13  FOR i60 in iS3{8}rf:
14    T2_s[ iS3{8}rf ]
   = pad( T1_l[ iS1{7} ], {0, 1} )
16  T3_l[ iS4{8} ] = ALLOCATE(buffer=T3_l[ iS4{8} ], mem_type=register, size=8, zero_init=false)
   ^^^^^ ---Buffer Reuse Info---   not aliased  , inner live interval: [ 18 , 21 ] , size expr : 8 , outer live interval: [ 17 , 22 ]    
17  FOR i61 in iS4{8}:
18    T3_l[ iS4{8} ]
   = -T2_s[ iS3{8}rf ];
20  FOR i62 in iS6{9}rf:
21    T4_s[ iS6{9}rf ]
   = pad( T3_l[ iS4{8} ], {0, 1} )
23  FOR i63 in iS7{9}:
24    T5_g[ iS7{9} ]
   = -T4_s[ iS6{9}rf ];
```
Notice that `T0_s` has an outer live interval with right endpoint at
position 12, which does not exist but signifies the end of the `FOR i59`
loop. In this case, `T0` was set for `promoteReuse` so a sync was
inserted in the interval `(12, 13)`.

Fixes #772.