Add SFINAE for the ctor of `DynamicType` (#792)
Just noticed that the casting operator from a custom class to
`DynamicType` is broken, the failing case is shown in
`DynamicTypeTest.CastToDynamicType`. The failure is because of the lack
of SFINAE in the ctor of `DynamicType`, when a conversion is needed, the
compiler always tries to use the ctor of `DynamicType` instead of the
casting operator. This PR adds the required SFINAE.