Ignore trivial loops in memory aliasing pass (#766)
Trivial `kir::ForLoop`s are ones that appear in the kernel IR, but do
not appear in the generated CUDA kernel. This can happen for a number of
reasons: for example if that dimension is vectorized, or if it's
parallelized with a stop value equal to the extent of a dimension. We
can test this with `kir::ForLoop::isTrivial()`. Consider an example:
```
T9_s[ ... ] ca_pos( 2 ) = ALLOCATE(buffer=T9_s[ ... ] ca_pos( 2 ), mem_type=shared, size=8192, zero_init=false)
T8_s[ ... ] ca_pos( 2 ) = ALLOCATE(buffer=T8_s[ ... ] ca_pos( 2 ), mem_type=shared, size=4096, zero_init=false)
T7_s[ ... ] ca_pos( 2 ) produce_pos( 2 ) = ALLOCATE(buffer=T7_s[ ... ] ca_pos( 2 ) produce_pos( 2 ), mem_type=shared, size=8192, zero_init=false)
FOR blockIdx.x in iblockIdx.x71{( ceilDiv(T0.logical_size[0], 64) )}:
  FOR blockIdx.y in iblockIdx.y73{( ceilDiv(T1.logical_size[1], 128) )}:
    ...
    FOR threadIdx.z in ithreadIdx.z84{( ceilDiv(( ( ceilDiv(64, 4) ) * 4 ), 32) )}:
      FOR threadIdx.y in ithreadIdx.y86{( ceilDiv(( ( ( ceilDiv(( ceilDiv(128, 8) ), 4) ) * 4 ) * 8 ), 32) )}:
        // All writes and reads of T7, T8, T9 are in this loop
        FOR i806 in iS88{( ceilDiv(32, 16) )}:
          T7_s = ...;
        ENDFOR i806
        T8_s = T7_s;
        T9_s = T8_s;
      ENDFOR threadIdx.y
    ENDFOR threadIdx.z
  ENDFOR blockIdx.y
ENDFOR blockIdx.x
```
In this case, all of the parallelized for loops are trivial, and only
the `FOR i806` loop appears in the generated code. That means the actual
lifetimes of T7 and T8 overlap and those of T8 and T9 overlap, but not
those of T7 and T9.

In the aliasing pass, we define outer live intervals as those at the
scope of the allocation. In the above case, it will set the outer live
interval of all three allocations equal to the start and end of the
`blockIdx.x` loop.

This PR ignores trivial loops in this analysis, so that outer live
intervals are defined at the scope _that will be realized in the CUDA
kernel_ at the level of the `Allocate` expression. In the above example,
this means the outer live intervals for T7 and T9 will no longer
overlap, so they are eligible for memory re-use.