Add `tools/compare_codegen.sh` (#787)
This is a development tool to help monitor the effect of code changes on
generated code. The tool accepts two optional arguments:
- The first argument is the git ref to compare against, which defaults
to `origin/main`.
- The second argument is the output directory holding all generated code
and log files, which defaults to `codegen_comparison`.

It requires the repository to be "clean" i.e. all changes committed. For
both the current commit and the reference commit, the following is done:
- Compiles code in `Release` mode
- Sets environment variables to ensure reproducibility and enable
dumping cuda code to files.
- Runs `nvfuser_tests` and `nvfuser_bench`, moving generated `.cu` files
to subdirectories named by git commit.
- Diffs the directories to find differences in generated code between
the two git commits.

If an error is encountered, or the script is interrupted with Ctrl-c,
then the original git head is restored.

If a directory exists for some commit, it is not recomputed. This lets
us maintain a common directory and compare to a single commit without
continually recomputing, if we wish.