Randomize testing with NVFUSER_TEST_RANDOM_SEED (#781)
There are two intentional sources of randomness in our C++ tests. These
are the ATen random factory functions like `at::randn`, and the C PRNG
accessed via `std::rand()`. The ATen randomness is used to create some
random data, and is compared with inexact matching at some tolerance:
these tests are expected to pass for most choices of random seed but not
necessarily for all of them, in order to balance false positive and
false negative rates. The C PRNG, on the other hand, is used for
structural randomness: choosing random sizes and indices of tensors when
creating a fusion. In these cases, we would ideally like to exhaustively
test all choices, but since that is impractical we use sampling instead.
Since we would like those cases to succeed for all choices of random
seed, we would like to inject more randomness. The intention of this PR
is to give us more control over random seeds in our tests and to enable
broader sampling of structural randomness using the C PRNG.

This PR changes `NVFuserTest` in the following ways:
- Sets ATen (using `at::manual_seed(seed)` and C PRNG seed (using
`std::srand(seed)`) in `SetUp()`. We no longer need to call `std::srand`
in tests, unless we want to explicitly use multiple seeds within one
test, such as in `FusionIndexSelectSimple_CUDA`.
- Adds the `NVFUSER_TEST_RANDOM_SEED` environment variable for setting
the C PRNG seed. This defaults to the time since the epoch in seconds.
- Adds the `NVFUSER_TEST_ATEN_RANDOM_SEED` environment variable for
setting the ATen random seed. This defaults to 0.
- Looks for failure in `TearDown()`. If test failed, we print the random
seeds used in the test so it can be reproduced.

If an error in a test occurs, we print the seeds that were used after
the backtrace. Sample output (truncated):

```
frame #14: <unknown function> + 0x29d90 (0x7fb716b71d90 in /usr/lib/x86_64-linux-gnu/libc.so.6)                                            
frame #15: __libc_start_main + 0x80 (0x7fb716b71e40 in /usr/lib/x86_64-linux-gnu/libc.so.6)                                                
frame #16: <unknown function> + 0x3fa725 (0x55b5ecb29725 in build/nvfuser_tests)                                                           
" thrown in the test body.                            
To reproduce: NVFUSER_TEST_RANDOM_SEED=1692968735 NVFUSER_TEST_ATEN_RANDOM_SEED=0 nvfuser_tests --gtest_filter='SmemReuseTest.PromoteReuseMultipleDownstream'
[  FAILED  ] SmemReuseTest.PromoteReuseMultipleDownstream (2009 ms)
```

This PR also changes `test_expr_simplifier.cpp` to use `std::rand()`
instead of hardware randomness. With this change, all tests in the test
suite are expected to be repeatable, when controlling the seeds as
described above.