Skip trivial Resize, handle resize of Broadcast at concretization. (#800)
This updates the root to rfactor propagation in IterType concretization
of dynamic fusions.

Previously, although we only overwrote Symbolic IterDomains in this
step, we still asserted that we could infer an IterType for each I moved
that check so that it is only applied when we need to make a change.

Additionally, we previously propagated Broadcast-only IterDomains as
Symbolic, since we combine with our previous estimate using
promoteIterType. As mentioned in a comment, this means Broadcast gets
propagated as Symbolic. Instead we now only fall back to promoteIterType
when there are multiple input IterTypes to the IterDomain expression.

Fixes #798