Enable shared memory reuse in matmul epilogue (#770)
This uses `promoteReuse` from #739 and inserts a syncthreads just before
the epilogue loop when smem is used for the epilogue, when possible. The
matmul heuristic attempts to predict when this will be possible in order
to more accurately estimate shared memory usage, and hence occupancy. If
we cannot guarantee re-use, we must assume in the heuristic that memory
will not be reclaimed, even though it might be when the fusion is
lowered.

Shared memory reclamation can only occur if the smem buffers have
non-overlapping lifetimes. This is difficult to guarantee before
scheduling and lowering. We use `cacheAfter` to create the `a` and `b`
smem tiles, but we use `cacheBefore` for the epilogue smem tile. This
means that smem will be used for any downstream uses of `a` and `b` but
the epilogue smem will have its lifetime restricted to the epilogue
itself, regardless of downstream uses of the matrix product.

The uses of `a` and `b` can complicate lifetime analysis. Consider a
case where both matrices are square and we wish to compute `a @ b + a`
where `@` denotes matmul. Since we used `a->cacheAfter()` to create the
smem tile, that smem may be used not only in the matmul but also in the
addition in the epilogue. In that case we cannot re-use a for the
epilogue smem. A conservative check that there are no other uses of `a`
or `b` is currently implemented in order to guarantee re-use. A less
conservative sufficient (but still not necessary) condition is that any
other use of `a` or `b` is a producer of `b` or `a` respectively; this
is not implemented yet.

---------

Co-authored-by: Andrzej Bekas <118676880+drzejan2@users.noreply.github.com>
Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>
Co-authored-by: Wang, Xiao <24860335+xwang233@users.noreply.github.com>