Enable object-based API option when compiling FlatBuffers schema (#812)
The PR enables the FlatBuffers object-based API. It allows for using
FlatBuffers objects directly in the NvFuser codebase and then
serializing those objects without any extra translation steps.

> --gen-object-api : Generate an additional object-based API. This API
is more convenient for object construction and mutation than the base
API, at the cost of efficiency (object allocation). Recommended only to
be used if other options are insufficient.

https://flatbuffers.dev/flatbuffers_guide_use_cpp.html

This option complements the existing API, so it can be used as-needed.
Normally, after deserialization, the FlatBuffer is deleted from memory,
so you lose access to the data. This option defines a cpp struct for
each table. You can then unpack data into a struct for persistent
storage.