[DynamicType] Fix automatic conversion to `DynamicType` (#850)
`PolymorphicValue->*something` returns an accessor of
`PolymorphicValue`, instead of returning just `PolymorphicValue`:

https://github.com/NVIDIA/Fuser/blob/0b85a578d968285f3e430c32de90708cc164bfda/csrc/polymorphic_value.h#L202

The accessor has defined automatic conversion operator to
`PolymorphicValue`, however, that automatic conversion does not work for
operators. For example:
```C++
PolymorphicValue x = ...;
PolymorphicValue y = ...;
x + (y->*"field");  // does not compile
```
This behavior does not make sense.

The root cause for this behavior is because, when we have
```C++
template<typename T>
T f(T x, T y) {
  ...
}
```
Then `f(2, 2L)` does not work, because `2` and `2L` are different type,
so there is no way for C++ to know whether it should choose `f<int>` or
`f<int64_t>`. For this case, automatic type conversion totally fails.
The correct way to implement `f` is to use a non-deductible environment
(the `type_identity` trick, see,
https://levelup.gitconnected.com/c-templates-type-identity-technique-f2b427759403):
```C++
template<typename T>
T f(T x, std::type_identity_t<T> y) {
  ...
}
```
This way, the type of `y` will not participate in template deduction,
and instead, it will be converted to whatever type x has. So `f(2, 2L)`
is equivalent to `f<int>(2, 2L)`. Always forcing the dtype of the first
arg is not the best behavior, but it does provide what we need: to make
accessors behave more like `PolymorphicValue` so it can be more
seamlessly used in our codebase.