Explicitly mutate either Expr inputs or outputs in OptOutMutator (#856)
Fixes #852.

This PR:
- Changes `OptOutMutator::mutate(Expr* expr)` such that registered
mutations of outputs of `expr` are not processed in that method. This
means that if we register a `Val` for mutation then call `mutate` only
on its definition, we will not wind up redefining that `Val` or its
replacement.
- Adds `OptOutMutator::mutateExprOutputsOnly(Expr* expr)`, which does
not replace any inputs or attributes to `expr`, but does replace outputs
to the expression with registered mutations. This has the effect of
transferring the definitions of those original outputs onto their
registered replacements.
- Removes `ir_utils::replaceValInExpr`.
- Adds `ir_utils::replaceValInExprInputs`, which works similarly to the
original `replaceValInExpr` but does not include the confusing
output-mutating behavior. Most uses of `replaceValInExpr` are converted
to use this directly.
- Adds `ir_utils::transferDefinitionToNewOutputs(Expr* expr, const
std::vector<Val*>& new_outputs)`. Like
`OptOutMutator::mutateExprOutputsOnly`, this does not replace inputs or
attributes of `expr`, but creates a new `Expr` which serves as the
definition of all members of `new_outputs`.

With these changes, registering a `Val` for mutation, then calling
`OptOutMutator::dispatchMutate(stmt)` on all `Statement`s in topological
order will produce the expected output and avoid strange cases like
those outlined in #852.

Note that there is still one location that depends on the original
behavior of `replaceValInExpr`: dynamic Fusion concretization.
Currently, when propagating concretized `IterDomain`s from root to
rfactor, we process `IterDomain` expressions in topological order then
call `mutate(expr)`, which alters not only the inputs of each expression
but also transfers that expression as the new definition for
previously-undefined `IterDomain`s that were registered as mutations of
the lhs of those expressions. In this PR, now we simply do this with two
separate function calls, explaining that we need to transfer the
expression as the new definition of the replacements, then we use the
new `Expr` created in that step and mutate it so that its inputs are
updated.