Tiny build changes (#896)
Even when BUILD_PYTHON doesn't get set, we still need to be able to
serialize things. Without this change we have undefined references when
BUILD_PYTHON is false or unset.

The second issue is that we couldn't actually set custom flags at cmake
invocation time, unless those flags happened to end in a separator.