Simplify more expressions in `SimplifyingIrBuilder` (#892)
This PR adds handling methods for comparison operators in
`SimplifyingIrBuilder`. It also adds `sameAs` checks to a few existing
methods to simplify expressions like `s0 - s0 = 0`, `ceilDiv(s1, s1) =
1`, or `s2 % s2 = 0`.

To implement the simplifying comparisons, we introduce
`csrc/C++20/compare` holding `partial_ordering` which is a class that
can represent a three-way comparison with an "unordered" possibility. In
this case, if either of the arguments to
`SimplifyingIrBuilder::ltExpr(Val* lhs, Val* rhs)` are non-constant, the
comparison is unordered and we return `IrBuilder::ltExpr(lhs, rhs)`.