Resolve broadcasts resulting from a PadOp (#610)
This changes broadcast analysis to drop the assumption that all
broadcasts are created through a `BroadcastOp`. This change allows a
broadcast to be introduced between root and rfactor of any tensor, which
happens if we pad an `Iteration` domain with negative pad widths,
resulting in a `Broadcast` domain.

This adds a `map_symbolic` option to `PairwiseRootDomainMap`, which
defaults to false and has the following effect (condition 5):
- When `map_symbolic == false`, symbolic `IterDomain`s are only mapped
with one another if their extent expressions match (`sameAs`). This is
the default mode and safely handles cases where one symbolic value will
concretize to broadcast that is resolved by the other.
- When `map_symbolic == true`, symbolic `IterDomain`s are mapped
regardless of their extent expressions. This mode is useful for producer
to consumer mappings when we know we are not resolving a broadcast, but
a placeholder extent expression might be used in the consumer. It is
only used for propagating concretizations from producer to consumer, in
which case we need to exact map aligned symbolic IDs.

Fixes #596. Fixes #357