Add diff_codegen_nvfuser_tests.py (#953)
This is just a small utility that will diff codegen output for
`nvfuser_tests` in a way that lines up output from the same tests, even
if new tests have been introduced.

This addresses a limitation of the default script where new tests lead
to more cuda kernels being output at some intermediate point in the run.
After that, all the kernels are misnumbered so we wind up with a huge
number of false positives.

This script parses the stdout log of each command to find which kernels
were produced by each test, and diffs those corresponding files
directly. The exit code is the number of tests with mismatches.

A typical usage would be something like the following:
```
> tools/compare_codegen.sh  # create codegen_comparison with subdirs for this commit and merge-base
> python tools/diff_codegen_nvfuser_tests.py codegen_comparison/*/binary_tests
```

Note that this tool can only be run with directories from
`nvfuser_tests` commands. Benchmarks and python tests are not supported.