Add LL suffix to Int literals (#998)
In https://github.com/NVIDIA/Fuser/issues/973#issuecomment-1741343225,
it was mentioned that expressions on `DataType::Int` constants are not
necessarily safe from overflow. For example, an `exp` expression on
`Int` constants currently prints like the example `2 ** 30`. Integer
literals that are not very large will be compiled as 32-bit ints, so
this arithmetic will be taking place at too low precision. This PR
addresses this issue by adding a suffix of `LL` for `Int` constants so
that these will translate to long long, which is 64-bit (see comment in
code and second reference below on why we don't use `L` instead).

References:
- https://en.cppreference.com/w/cpp/language/integer_literal
- https://en.cppreference.com/w/cpp/language/types