Fill the final loop promotion map  (#1028)
This PR adds structural validations of loop graphs and promotions of one
of the most important tests, `Indexing19`. In order to do so, it
introduces a minimal set of changes as follows:

- Enable building loop graphs and promotion mappings even for `Fusion`
- Populate a map from loop ID groups to their promotion domains, which
is temporarily done at the very end of `buildLoopPromotionMap`.

Some cleanups are needed, but I will iteratively work on cleanups as
well as functional changes. This PR severs as the baseline for
forthcoming PRs as well as making sure my understanding is correct.