[initial build up] `mbarrier`: arrive wait barrier on smem (#995)
Fixes: https://github.com/NVIDIA/Fuser/issues/992 Required by:
https://github.com/NVIDIA/Fuser/issues/993

This PR introduces `mbarrier`, an arrive-wait barrier on shared memory.
The code for `mbarrier` itself is ready-to-use, however, there is no
passes in our lowering currently using this barrier. In future PR, I
will explore changing our block syncs with `mbarrier` when makes sense.

In this PR, a new test `MBarrierTest.Simple` is added. This test is a
simple `gmem->smem->gmem` copy kernel. The fusion is scheduled in a way
that block sync is needed. And in the test, the lowered kernel is
modified to replace the block sync with `mbarrier`. Because there is no
lowering pass using `mbarrier`, this test is written in a hacky way that
it lowers to a kernel first and then modifies the lowered kernel.