Fix mapping propagations through uses
This doens't seem necessary at this moment but required to stop using
ALMOST_EXACT for permissive mappings.

Suppose domains A and B have uses as:

A: ExprGroup X and ExprGroup Y
B: ExprGroup X and ExprGroup Y

i.e., both have the same set of expr groups. Before this propagation,
there's nothing to allow map X and Y. When A and B are mapped, we should
map the uses of A and B, which means X and Y should be mapped. The check
removed in this PR prevents the propagation.